connect4/requirements.txt
Pillow
numpy
PyYAML
scipy
wheel
https://github.com/pybrain/pybrain/archive/0.3.3.zip
